Fixed react/jsx-filename-extension. Additionally renamed directory "General" to "general"
