+++ frontend/src/StatusIndicator.jsx
