+++ frontend/src/createStream/SelectDeviceTableRow.jsx
