+++ frontend/src/devicelist/ChannelDetailsTable.jsx
