+++ frontend/src/devicelist/ContentsTable.jsx
