+++ frontend/src/devicelist/SearchBar.jsx
