+++ frontend/src/devicelist/SingleTableRow.jsx
