+++ frontend/src/devicelist/TitleBox.jsx
@@ -2,8 +2,8 @@ import React from "react";
 import { Box } from "@material-ui/core";
 import { NavLink } from "react-router-dom";
 
-import StreamButton from "../General/Buttons/StreamButton";
-import AddDeviceButton from "../General/Buttons/AddDeviceButton";
+import StreamButton from "../general/Buttons/StreamButton";
+import AddDeviceButton from "../general/Buttons/AddDeviceButton";
 
 export default class TitleBox extends React.Component {
   render() {
