+++ frontend/src/devicelist/VerticalTabs.jsx
