+++ frontend/src/general/Buttons/AddDeviceButton.jsx
